requirements.txt
pytest==8.1.1
pylint==3.1.0
selenium==4.20.0
webdriver-manager==4.0.1
psycopg2==2.9.9
SQLAlchemy==1.4.50
allure-pytest==2.13.5
requests==2.31.0
